requirements.txt
ipykernel==6.29.4
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.4
langchain_core
langchain_text_splitters
openai==1.24.0
streamlit==1.35.0
python-dotenv==1.0.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.23.1